--- src/components/FormContainer/FormContainer.stories.tsx
@@ -1,23 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import FormContainer from "./FormContainer";
-import { useState } from "react";
-
-const meta = {
-  title: "MyComponent/MyFormContainer",
-  component: FormContainer,
-  tags: ["autodocs"],
-} satisfies Meta<typeof FormContainer>;
-
-export default meta;
-
-type Story = StoryObj<typeof FormContainer>;
-
-export const Primary: Story = {
-  args: {
-    cardState: { uniqueNumber: ["", "", "", ""], expirationDate: ["", ""], cvcNumber: [""] },
-    dispatch: () => {},
-  },
-  render: (args: any) => {
-    return <FormContainer cardState={args.cardState} dispatch={args.dispatch} />;
-  },
-};
