+++ src/components/feature/PreviewCard/PreviewCard.tsx
@@ -12,7 +12,6 @@ const PreviewCard = ({ uniqueNumber, expirationDate }: { uniqueNumber: string[];
     if (/^5[1-5]/.test(bin)) return "master";
     return "none";
   }
-  //utils로 분리
 
   const cardType = getCardType(uniqueNumber[0]);
