+++ vite.config.ts
@@ -1,11 +1,10 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
 
-// https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
     react({
-      jsxImportSource: "@emotion/react", // ✅ 이것만 있어도 기본적으로 작동함
+      jsxImportSource: "@emotion/react",
     }),
   ],
 });
